+++ gantry/__main__.py
@@ -167,8 +167,8 @@ def main():
     "--dataset",
     type=str,
     multiple=True,
-    help="""An input dataset in the form of 'dataset-name:/mount/location' to attach to your experiment.
-    You can specify this option more than once to attach multiple datasets.""",
+    help="""An input dataset in the form of 'dataset-name:/mount/location' or 'dataset-name:/subpath:/mount/location'
+    to attach to your experiment. You can specify this option more than once to attach multiple datasets.""",
 )
 @click.option(
     "--gh-token-secret",